Render agent results directly in React
