--- apps/template/lib/agent/index.ts
@@ -1,47 +1,0 @@
-import { defineCatalog } from "@json-render/core";
-import { schema } from "@json-render/react/schema";
-import { z } from "zod";
-
-export const catalog = defineCatalog(schema, {
-  actions: {},
-  components: {
-    AgentResponse: {
-      description: "A stack of storefront-rendered shopping results.",
-      props: z.object({}),
-      slots: ["default"],
-    },
-    AgentCartSummary: {
-      description:
-        "The shopper's live cart with quantity steppers, remove buttons, totals, and a checkout button. " +
-        "Takes no props — it reads real cart state. Render it for get-cart, not cart mutations.",
-      props: z.object({}),
-    },
-
-    AgentProductCard: {
-      description:
-        "A single product card. Pass only the handle returned by a product tool; the card resolves " +
-        "its own image, title, price, and availability from that tool result.",
-      props: z.object({
-        handle: z.string(),
-      }),
-    },
-
-    AgentProductGrid: {
-      description:
-        "A responsive grid container for AgentProductCard children. Always wrap multiple product cards in this grid.",
-      props: z.object({
-        title: z.string().nullable(),
-      }),
-      slots: ["default"],
-    },
-
-    AgentVariantPicker: {
-      description:
-        "An interactive variant picker for one product. Pass only the handle from getProductDetails. " +
-        "The shopper selects options and adds to cart directly, so do not ask them to type a variant choice.",
-      props: z.object({
-        handle: z.string(),
-      }),
-    },
-  },
-});
